+++ src/runner/v2/platform/assets/NullVideoAsset.ts
@@ -11,7 +11,10 @@ export class NullVideoAsset extends g.VideoAsset {
 	}
 
 	_load(loader: g.AssetLoadHandler): void {
-		loader._onAssetLoad(this);
+		// アセットのロードは原則非同期としたいのでsetTimeoutを使用
+		setTimeout(() => {
+			loader._onAssetLoad(this);
+		}, 0);
 	}
 
 	asSurface(): g.Surface {