+++ src/runner/v3/platform/graphics/null/NullImageAsset.ts
@@ -17,7 +17,10 @@ export class NullImageAsset extends Asset implements g.ImageAsset {
 	}
 
 	_load(loader: g.AssetLoadHandler): void {
-		loader._onAssetLoad(this);
+		// アセットのロードは原則非同期としたいのでsetTimeoutを使用
+		setTimeout(() => {
+			loader._onAssetLoad(this);
+		}, 0);
 	}
 
 	asSurface(): g.Surface {